+++ src/G2oTypes.cc
@@ -158,7 +158,7 @@ void ImuCamPose::SetParam(const std::vector<Eigen::Matrix3d>& _Rcw,
   Rcb.resize(num_cams);
   tcb.resize(num_cams);
 
-  for (int i = 0; i < tcb.size(); i++) {
+  for (size_t i = 0; i < tcb.size(); i++) {
     Rcb[i] = Rbc[i].transpose();
     tcb[i] = -Rcb[i] * tbc[i];
   }
@@ -209,7 +209,7 @@ void ImuCamPose::Update(const double* pu) {
   const Eigen::Matrix3d Rbw = Rwb.transpose();
   const Eigen::Vector3d tbw = -Rbw * twb;
 
-  for (int i = 0; i < pCamera.size(); i++) {
+  for (size_t i = 0; i < pCamera.size(); i++) {
     Rcw[i] = Rcb[i] * Rbw;
     tcw[i] = Rcb[i] * tbw + tcb[i];
   }
@@ -242,7 +242,7 @@ void ImuCamPose::UpdateW(const double* pu) {
   const Eigen::Matrix3d Rbw = Rwb.transpose();
   const Eigen::Vector3d tbw = -Rbw * twb;
 
-  for (int i = 0; i < pCamera.size(); i++) {
+  for (size_t i = 0; i < pCamera.size(); i++) {
     Rcw[i] = Rcb[i] * Rbw;
     tcw[i] = Rcb[i] * tbw + tcb[i];
   }
@@ -294,7 +294,7 @@ bool VertexPose::write(std::ostream& os) const {
 
   const size_t num_cams = tcw.size();
 
-  for (int idx = 0; idx < num_cams; idx++) {
+  for (size_t idx = 0; idx < num_cams; idx++) {
     for (int i = 0; i < 3; i++) {
       for (int j = 0; j < 3; j++) os << Rcw[idx](i, j) << " ";
     }
